👇 Floor #6 - Add Carpet and Hardwood for Rooms
